fix(UI): 에디터에서 버전 라벨이 git 태그를 반영하도록 수정
VersionStamp는 빌드 시에만 bundleVersion을 각인해, 에디터에서는
Application.version이 기본값에 머물러 버전 라벨이 실제 git 태그와
달랐다. git 해석 로직을 GitVersion 헬퍼로 분리하고, 에디터에서는
VersionLabel이 git 태그를 직접 읽어 표시하도록 한다(표시용 메타데이터
+dirty/+N.g 접미사는 제거). bundleVersion은 건드리지 않는다.
