+++ ChaosChess_v2/Assets/Script/UI/GitVersion.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 3fcd83d49b0b3144eb3a9ba02493ec10
